Completed Api code
In this code contains full functionality api codes

+++ controllers/category.controller.js
@@ -1,15 +1,19 @@
 const Category = require('../models/category.model');
 const mongoose = require('mongoose');
-const CustomObjectId = mongoose.Types.ObjectId;
-
+// const CustomObjectId = mongoose.Types.ObjectId;
+const subcategory = require('../models/subcategory.model');
+const expenditures = require('../models/expenditures.model');
 
 exports.addCategory = async (req, res) => {
     try {
+        // console.log("getting resonse from middlewere",req.user)
+        // res.send(req.user);
+        // return;
         const cgry = new Category({
             title: req.body.title,
-            userId : req.body.userId
+            userId: req.user._id
         })
-       
+
         cgry.save().then(x => {
             return res.status(200).json({
                 "status": "200",
@@ -30,15 +34,15 @@ exports.addCategory = async (req, res) => {
 }
 exports.getAllCategory = async (req, res) => {
     try {
-        Category.find({userId : req.query.userId})
-        .then(x => {
-            return res.status(200).json({
-                "status": "200",
-                "success": true,
-                "response" : x,
-                "msg": "Gets all category successfully!!!",
+        Category.find({ userId: req.user._id })
+            .then(x => {
+                return res.status(200).json({
+                    "status": "200",
+                    "success": true,
+                    "response": x,
+                    "msg": "Gets all category successfully!!!",
+                })
             })
-        })
     } catch (err) {
         return res.status(422).json({
             "statusCode": "422",
@@ -51,15 +55,15 @@ exports.getAllCategory = async (req, res) => {
 }
 exports.updateCategory = async (req, res) => {
     try {
-        Category.updateOne({userId : req.body.userId,_id : req.body.categoryId },{$set : {title : req.body.title}})
-        .then(x => {
-            return res.status(200).json({
-                "status": "200",
-                "success": true,
-                res : x,
-                "msg": "Category Updated successfully!!!",
+        Category.updateOne({ userId: req.user._id, _id: req.body.categoryId }, { $set: { title: req.body.title,updateddAt : Date.now() } })
+            .then(x => {
+                return res.status(200).json({
+                    "status": "200",
+                    "success": true,
+                    res: x,
+                    "msg": "Category Updated successfully!!!",
+                })
             })
-        })
     } catch (err) {
         return res.status(422).json({
             "statusCode": "422",
@@ -72,13 +76,25 @@ exports.updateCategory = async (req, res) => {
 }
 exports.deleteCategory = async (req, res) => {
     try {
-        Category.findByIdAndDelete({_id : req.params.categoryId}).then(x => {
+        const subcategoryData = await subcategory.find({ categoryId: req.params.categoryId })
+        const expen = await expenditures.find({ categoryId: req.params.categoryId })
+        if (subcategoryData.length > 0 || expen.length > 0) {
             return res.status(200).json({
                 "status": "200",
                 "success": true,
-                "msg": "Category Deleted successfully!!!",
+                "res": 0,
+                "msg": "Category contains subcategory datas!!!",
             })
-        })
+        } else {
+            Category.findByIdAndDelete({ _id: req.params.categoryId }).then(x => {
+                return res.status(200).json({
+                    "status": "200",
+                    "success": true,
+                    "res": 1,
+                    "msg": "Category Deleted successfully!!!",
+                })
+            })
+        }
     } catch (err) {
         return res.status(422).json({
             "statusCode": "422",
