+++ controllers/paymnet.controller.js
@@ -1,9 +1,10 @@
 const payment = require('../models/paymenttype.model');
+const expenses = require('../models/expenses.model');
 exports.addPayment = async (req, res) => {
     try {
         const pm = new payment({
             title : req.body.title,
-            userId : req.body.userId
+            userId : req.user._id
         })
         pm.save()
         .then( x => {
@@ -25,7 +26,7 @@ exports.addPayment = async (req, res) => {
 }
 exports.getAllPayment = async (req, res) => {
     try {
-        payment.find({userId : req.query.userId})
+        payment.find({userId : req.user._id})
         .then( x => {
             return res.status(200).json({
                 "status": "200",
@@ -46,7 +47,7 @@ exports.getAllPayment = async (req, res) => {
 }
 exports.updatePayment = async (req, res) => {
     try {
-        payment.updateOne({_id : req.body.paymentId},{$set : {title : req.body.title}})
+        payment.updateOne({_id : req.body.paymentId},{$set : {title : req.body.title,updatedAt : Date.now()}})
         .then( x => {
             return res.status(200).json({
                 "status": "200",
@@ -66,14 +67,25 @@ exports.updatePayment = async (req, res) => {
 }
 exports.deletePayment = async (req, res) => {
     try {
-        payment.findByIdAndDelete({_id : req.params.paymentId})
-        .then( x => {
+        const expen = await expenses.find({ paymentType: req.params.paymentId });
+        if(expen.length > 0){
             return res.status(200).json({
                 "status": "200",
                 "success": true,
+                "res" : 0,
                 "msg": "Payment Deleted successfully!!!",
             })
-        })
+        }else{
+            payment.findByIdAndDelete({_id : req.params.paymentId})
+            .then( x => {
+                return res.status(200).json({
+                    "status": "200",
+                    "success": true,
+                    "res" : 1,
+                    "msg": "Payment Deleted successfully!!!",
+                })
+            })
+        }
     } catch (err) {
         return res.status(422).json({
             "statusCode": "422",
